feat: remove ktfmt in favour of ktlint and detekt analyser

+++ src/main/kotlin/uk/me/danielharman/kotlinspringbot/services/SpringGuildService.kt
@@ -1,6 +1,5 @@
 package uk.me.danielharman.kotlinspringbot.services
 
-import kotlin.math.max
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import org.springframework.data.domain.PageRequest
@@ -10,10 +9,12 @@ import uk.me.danielharman.kotlinspringbot.helpers.OperationResult
 import uk.me.danielharman.kotlinspringbot.helpers.Success
 import uk.me.danielharman.kotlinspringbot.models.SpringGuild
 import uk.me.danielharman.kotlinspringbot.repositories.GuildRepository
+import kotlin.math.max
 
 @Service
-class SpringGuildService(private val guildRepository: GuildRepository) {
-
+class SpringGuildService(
+    private val guildRepository: GuildRepository,
+) {
     private val logger: Logger = LoggerFactory.getLogger(this::class.java)
 
     fun getGuild(guildId: String): OperationResult<SpringGuild, String> {
@@ -37,23 +38,25 @@ class SpringGuildService(private val guildRepository: GuildRepository) {
         return Success(guildRepository.save(SpringGuild(guildId)))
     }
 
-    fun deleteSpringGuild(guildId: String): OperationResult<String, String> {
-        return when (val guild = getGuild(guildId)) {
+    fun deleteSpringGuild(guildId: String): OperationResult<String, String> =
+        when (val guild = getGuild(guildId)) {
             is Failure -> guild
             is Success -> {
                 guildRepository.deleteByGuildId(guild.value.guildId)
                 Success(guild.value.id)
             }
         }
-    }
 
-    fun getGuilds(pageSize: Int = 10, page: Int = 0): OperationResult<List<SpringGuild>, String> =
+    fun getGuilds(
+        pageSize: Int = 10,
+        page: Int = 0,
+    ): OperationResult<List<SpringGuild>, String> =
         Success(guildRepository.findAll(PageRequest.of(max(page, 0), max(pageSize, 1))).toList())
 
     fun updateUserCount(
         guildId: String,
         userId: String,
-        count: Int
+        count: Int,
     ): OperationResult<SpringGuild, String> {
         val guild = createGuildIfNotExists(guildId)
         if (guild is Failure) return guild
@@ -63,7 +66,10 @@ class SpringGuildService(private val guildRepository: GuildRepository) {
         return Success(result)
     }
 
-    fun setVol(guildId: String, vol: Int): OperationResult<Int, String> {
+    fun setVol(
+        guildId: String,
+        vol: Int,
+    ): OperationResult<Int, String> {
         val guild = createGuildIfNotExists(guildId)
         if (guild is Failure) return guild
 
@@ -81,15 +87,17 @@ class SpringGuildService(private val guildRepository: GuildRepository) {
         return Success(findAndModify.volume)
     }
 
-    fun getVol(guildId: String): OperationResult<Int, String> {
-        return when (val guild = getGuild(guildId)) {
+    fun getVol(guildId: String): OperationResult<Int, String> =
+        when (val guild = getGuild(guildId)) {
             is Failure -> guild
             is Success -> Success(guild.value.volume)
         }
-    }
 
-    fun isModerator(guildId: String, userId: String): OperationResult<String, String> {
-        return when (val guild = getGuild(guildId)) {
+    fun isModerator(
+        guildId: String,
+        userId: String,
+    ): OperationResult<String, String> =
+        when (val guild = getGuild(guildId)) {
             is Failure -> guild
             is Success -> {
                 if (guild.value.privilegedUsers.contains(userId)) {
@@ -99,9 +107,11 @@ class SpringGuildService(private val guildRepository: GuildRepository) {
                 }
             }
         }
-    }
 
-    fun addModerator(guildId: String, userId: String): OperationResult<String, String> {
+    fun addModerator(
+        guildId: String,
+        userId: String,
+    ): OperationResult<String, String> {
         return when (val guild = createGuildIfNotExists(guildId)) {
             is Failure -> guild
             is Success -> {
@@ -112,7 +122,10 @@ class SpringGuildService(private val guildRepository: GuildRepository) {
         }
     }
 
-    fun removeModerator(guildId: String, userId: String): OperationResult<String, String> {
+    fun removeModerator(
+        guildId: String,
+        userId: String,
+    ): OperationResult<String, String> {
         when (val guild = getGuild(guildId)) {
             is Failure -> return guild
             is Success -> {
@@ -123,7 +136,10 @@ class SpringGuildService(private val guildRepository: GuildRepository) {
         }
     }
 
-    fun addMemeChannel(guildId: String, channelId: String): OperationResult<String, String> {
+    fun addMemeChannel(
+        guildId: String,
+        channelId: String,
+    ): OperationResult<String, String> {
         return when (val guild = createGuildIfNotExists(guildId)) {
             is Failure -> guild
             is Success -> {
@@ -134,7 +150,10 @@ class SpringGuildService(private val guildRepository: GuildRepository) {
         }
     }
 
-    fun removeMemeChannel(guildId: String, channelId: String): OperationResult<String, String> {
+    fun removeMemeChannel(
+        guildId: String,
+        channelId: String,
+    ): OperationResult<String, String> {
         return when (val guild = getGuild(guildId)) {
             is Failure -> guild
             is Success -> {
@@ -145,14 +164,16 @@ class SpringGuildService(private val guildRepository: GuildRepository) {
         }
     }
 
-    fun getMemeChannels(guildId: String): OperationResult<List<String>, String> {
-        return when (val guild = getGuild(guildId)) {
+    fun getMemeChannels(guildId: String): OperationResult<List<String>, String> =
+        when (val guild = getGuild(guildId)) {
             is Failure -> guild
             is Success -> Success(guild.value.memeChannels)
         }
-    }
 
-    fun deafenChannel(guildId: String, channelId: String): OperationResult<String, String> {
+    fun deafenChannel(
+        guildId: String,
+        channelId: String,
+    ): OperationResult<String, String> {
         return when (val guild = getGuild(guildId)) {
             is Failure -> guild
             is Success -> {
@@ -163,7 +184,10 @@ class SpringGuildService(private val guildRepository: GuildRepository) {
         }
     }
 
-    fun unDeafenChannel(guildId: String, channelId: String): OperationResult<String, String> {
+    fun unDeafenChannel(
+        guildId: String,
+        channelId: String,
+    ): OperationResult<String, String> {
         return when (val guild = getGuild(guildId)) {
             is Failure -> guild
             is Success -> {
@@ -174,20 +198,20 @@ class SpringGuildService(private val guildRepository: GuildRepository) {
         }
     }
 
-    fun isChannelDeafened(guildId: String, channelId: String): Boolean {
-        return when (val guild = getGuild(guildId)) {
+    fun isChannelDeafened(
+        guildId: String,
+        channelId: String,
+    ): Boolean =
+        when (val guild = getGuild(guildId)) {
             is Failure -> false
             is Success -> guild.value.deafenedChannels.contains(channelId)
         }
-    }
 
-    fun getDeafenedChannels(guildId: String): OperationResult<List<String>, String> {
-        return when (val guild = getGuild(guildId)) {
+    fun getDeafenedChannels(guildId: String): OperationResult<List<String>, String> =
+        when (val guild = getGuild(guildId)) {
             is Failure -> guild
             is Success -> Success(guild.value.deafenedChannels)
         }
-    }
 
-    fun getGuildsWithoutModerators(): OperationResult<List<SpringGuild>, String> =
-        Success(guildRepository.getGuildsWithModeratorCount(0))
+    fun getGuildsWithoutModerators(): OperationResult<List<SpringGuild>, String> = Success(guildRepository.getGuildsWithModeratorCount(0))
 }
